+++ src/core/layers/index.ts
@@ -56,5 +56,6 @@ export {
 
 export {
   default as YAxisLayer,
-  Props as YAxisLayerProps
+  Props as YAxisLayerProps,
+  YAxisControls
 } from './YAxisLayer';